MB-60680 Allow non-equal predicate on leading scalar index keys ...    ... for ORDER and LIMIT pushdown in case of vector index
Normally for index ordering as soon as we encounter a non-equal
predicate we bail out since order can no longer be satisfied by
the index scan. For ordering of vector predicate though, since
the indexer uses a max heap to get the ordering for vector
distances, we can allow non-equal predicate on leading scalar
index keys to satisfy index order. In fact even trailing scalar
index keys can be ordered as well. It does require however that
the index scan is "exact", and order by vector predicate is
required. This also allows the situation where we assume LIMIT
is pushed down (which is required for max heap to work), but
for whatever reason the LIMIT cannot (which is determined after
ORDER pushdown), we'll then go through this function again
with ordering of vector index key disabled, which will then
turn off ORDER pushdown for the vector index key (this part of
the logic already exists).

Change-Id: I2345456f063a78c87624601e2df0bc35034b7966
Reviewed-on: https://review.couchbase.org/c/query/+/217457
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>