MB-62471: [4/n] AUS task cancellation and timeout
Running AUS tasks that are deleted or the scheduled window of
the task

Running AUS tasks can be cancelled in the following ways:
1. The scheduled window of the task is exceeded.
2. The task is cancelled i.e deleted from system:tasks_cache

For both cases, when the AUS task is notified to stop its execution
any currently executing Update Statistics is cancelled. And further
processing of the AUS task is prevented.

Scheduler changes:
The current implementation of the scheduler does not allow running
tasks to be aborted. Once the task execution begins, any attempt
to delete it, is not allowed.

To accomadate the need for AUS tasks to be cancelled the following
changes have been made to the scheduler:

* When a task is scheduled, it can opt-in to allow the task
to be "stoppable" i.e its execution aborted once begun.

* This behavior is provided by the introduction of a new scheduler
function - ScheduleStoppableTask()

* The execution function of such a task, is notified via a channel
if its execution is to be stopped. And must do the appropriate
actions to stop its actions once notified.

New scheduler task states introduced:
1. ABORTING - a running task’s execution is in the process of
              being aborted
2. ABORTED - a running task’s execution was aborted

Note - this "stoppable" nature of the task execution is opt-in.
Hence, if this behaviour is not desired, the existing scheduling
function i.e ScheduleFunction() can be used.

Hence, the behaviour of other tasks that use the scheduler like
ADVISOR sessions, UPDATE STATISTICS statements is not changed.

* Also made changes to the scheduler to prevent checking and changing
task state, performing cleanups without synchronisation.

Change-Id: Ifdf10c20b5bcc8abfba1c98e31d98a8173293ba5
Reviewed-on: https://review.couchbase.org/c/query/+/217758
Well-Formed: Restriction Checker
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>