MB-63659 Add SQL-like syntax support for ...
... natural language requests.

This adds:

USING AI [WITH {...}] <prompt>

as a valid form for the "statement" request parameter.  The options
specified in the WITH clause as processed into the appropriate
"natural_..." parameters and the <prompt> into the "natural" parameter
(with "statement" being unset).  From this point on it is then treated
identically to a "regular" natural language request.

Options supported in the WITH clause are:

"creds" - Either in "natural_cred" form (user:password) or in common
          JSON form (an object with members "user" and "pass").
          Populates "natural_cred".
"orgid" - Populates "natural_orgid".
"keyspaces" - Either a single string or an array of strings.  The
              strings are joined into a single comma-delimited list.
              Populates "natural_context".

If the appropriate "natural_..." options are supplied the WITH options
aren't required, but the WITH options will take precedence (overwrite
the request parameters).

In addition, natural language processing errors have been revised so
they may also make sense when no "natural_..." request parameters are
involved.

Revises processing of "natural_context".  Moves the processing into a
generic "parse and validate a list of paths" function which also permits
optional element quoting and is tolerant of spaces between elements and
delimiters.

Revises the prompt so that the LLM's missing keyspace information
message is shown as it is clearer, the explanation for errors is devoid
of a prefix (just makes the errors look better), aliases used in
statements are quoted (so aliases that collide with reserved words
aren't an issue), require the generation of a single statement for the
prompt (this results in a UNION ALL query when for example you list
multiple keyspaces and prompt with "list all documents"; previously
it'd generate one statement per keyspace which we'd then fail to
process) and require the statement be generated on a single line (so
that cut-n-paste of the "generated_statement" element is simpler when
it is necessary/desired as there are no JSON-escaped newline characters
to deal with).

Error examples (to illustrate the changes):

"msg": "LLM error: The prompt \"blah\" does not provide sufficient
information to generate a specific SQL++ query."

"msg": "LLM error: The provided information about keyspaces does not
include any details about an \"employees\" collection or related schema.
Therefore, it is not possible to generate a valid query to list all
employees based on the given information."

Addes a metric for the "naturalLanguageProcessingTime" that reports the
time spent in the natural language processing.  (Also present in
completed_requests.)  This processing time is marked by a new request
state "pre-processing" as this can take some time and is then visible in
active_requests (easy to identify a request which is busy in the natural
language processing).

Adds specific handling for "natural_cred" in cbq-shell: when the value
supplied for it in a \SET command lacks the password delimiter (:) it
prompts via the terminal for the password.  This provides a means to
keep the password from shared access (like shell or cbq-shell history).

Also reformats the rest of the messages.go to conform to line-length
limits.

Change-Id: Icfed217a8748dbc3e208d3156b069cb601da204b
Reviewed-on: https://review.couchbase.org/c/query/+/217824
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>