MB-63918 Improve arrayId assignment for MERGE ON clause
This bug is similar to MB-63163, and is a regression from MB-61549.
The fix for MB-61549 introduced an arrayId for ANY predicates, and
the arrayId should be assigned early in the planner before any
potential index selection. The fix for MB-63163 already introduced
the assignment of arrayId for ON-clause of a MERGE statement,
however the issue is the assignment is done after the ON-clause is
being processed, during which copies of the ANY clause are already
made with the unassigned arrayId value of 0. The subsequent
assignment of arrayId only occurs on the original expression, not
the copies already made, and thus we still have expressions with
unassigned arrayId, causing issues later.

The fix is to move the assignment of arrayId earlier, before any
processing of the ON-clause predicates.

Change-Id: I62422ad5e5ec1da540e85cb41ff4a33107ba3a48
Reviewed-on: https://review.couchbase.org/c/query/+/218094
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/218103