MB-63921 Don't cache aggregate expression text
This change reinstates prior behaviour; commit
0c3dcd9a93352f5fb59fedb02247c1e53878d678 added caching of the text in an
effort to reduce load but doing so exposed this problem.  This is a
minimal-ish approach change to revert to the prior behaviour whilst
still attempting to address the memory aspect of the exposed issue and
better handle faults.

When an aggregate sub-query expression is evaluated, the expression may
be altered by the planner.  This means that agg.String(), used to key
the "aggregates" attachment, changes.  If this occurs during processing
the "aggregates" attachment for a grouping key needs to have the new key
inserted else subsequent passes through the grouping will fail as the
(new, updated) key doesn't exist.  This is the prior behaviour.
Previously we leaked the original "aggregates" entry; this change
removes the old entry when it detects the key has changed.

This still leaves some problems though:

1. If the expression changes again on some subsequent evaluation (e.g.
   an index is dropped whilst the group by is processing and the planner
   marks different parts as covered) only the grouping key that
   experienced the change whilst processing CumulateInitial will have
   the new "aggregates" key value.  All other group keys will have the
   prior value - and will panic the next time they're accessed.

2. agg.String() is potentially expensive both in terms of CPU and in
   terms of GC load through transient objects used for its execution.
   Repeating this twice for every aggregate, for every document is
   wasteful.

Change-Id: I912d537bc6c8341b3a66286fa5d3553202696961
Reviewed-on: https://review.couchbase.org/c/query/+/218211
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>