MB-63947 Modify how SubqueryTerm is marked during join enumeration
Currently join enumeration is marked by a flag on the builder, and
this flag has a couple of purposes:
  - to mark the actual process of join enumeration
  - if there is a SubqueryTerm in the query, when the SubqueryTerm
    is being planned as part of join enumeration, we currently also
    inherit this flag, which means during the entire planning of
    the SubqueryTerm, it is marked as join enumeration

The issue with the second part above is that, the planning of the
SubqueryTerm may or may not actually use join enumeration. In this
bug join enumeration is not used, however due to the presence of the
inherited flag, we are making certain decisions assuming we are
doing join enumeration, which can be problematic. There are a couple
of things that went wrong with this bug when the SubqueryTerm is
planned:
  - hash join is not being considered
  - with an outer join, we should process the ON clause when the
    outer join is being planned, however this is not done properly
    due to the presence of the join enumeration flag

The knowledge of a SubqueryTerm being planned as part of join
enumeration is required for the fix for MB-63702. To address the
issue for this bug while still maintaining the fix for MB-63702,
the following changes are done:
  - do not inherit the join enumeration flag when plannning a
    SubqueryTerm (as part of join enumeration)
  - instead, a new builder flag is introduced, to mark that a
    SubqueryTerm is being planned as part of join enumeration.
    This flag is somewhat similar to the other flag introduced as
    part of the fix for MB-63702, for marking a SubqueryTerm being
    planned under a nested-loop join
  - the new builder flag is inherited when planning the
    SubqueryTerm, instead of the join enumeration flag
  - modify the condition to also include the new builder flag when
    determining whether cover transformation should be skipped
  - include the error for primary index scan exceeding the limit
    in the errors to be ignored during join enumeration

A separate issue was identified during investigation, in that the
fix for MB-63702 does not handle nested SubqueryTerms when
performing covering transformation. Thus also add new logic to keep
track of nested SubqueryTerms, and also do cover transformation in
nested SubqueryTerm when cover transformation is done at the end of
join enumeration.

Also modified unit test infrastructure to ignore more log messages
from indexer client in the query log, such that in case of a test
failure it is easier to diagnose.

Change-Id: I49a723794e3719351fbb19dbabde7ee5ce67d819
Reviewed-on: https://review.couchbase.org/c/query/+/218307
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>