MB-63947 Make sure document count is available
This is an additional change for Morpheus only.

The reason the SubqueryTerm is not using join enumeration in this bug
is due to lack of document count for the keyspaces involved. The
join enumeration code first checks that document count is available
for all keyspaces, which is a prerequisite for cost calculations.
Currently document count is populated when predicates are processed
in the planning process. In case of the SubqueryTerm we only have a
LEFT JOIN between two keyspaces with no WHERE clause, as a result
the code for populating document count is not triggered.

Move the logic for populating document count to when baseKeyspace is
first constructed. For SELECT this is done when keyspace references
are gathered, for MERGE or UPDATE/DELETE this is done early on in
the planning process.

Change-Id: Id6161ec7964e9059ea716c61814b5dafc252c9a9
Reviewed-on: https://review.couchbase.org/c/query/+/218396
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>