MB-41307 Revise expression.Stringer...
... to not require as many temporary objects.  This both boosts straight
performance and reduces GC load.

The idea is the Stringer manages a single buffer which all shared visits
(and direct string writes) write to, with only a non-shared visit
consuming the buffer to return.  The buffer becomes the underlying
storage for the returned string so is not (must not be) used after this.
(It deliberately delays allocation again until a further write so in
cases where the stringer is used inline - though these have mainly been
removed - we don't have a second unused allocation.)

Users of the Stringer have also been revised to take advantage of being
able to share or have been converted to use the expression's String()
function directly, which (has always) just wraps an in-line Stringer.

Change-Id: I24604395c64d25ff46e2f96ffcad1dac6b24f3c5
Reviewed-on: https://review.couchbase.org/c/query/+/218715
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>