MB-64031 Improve SubsetOf functionality
When determining whether an expression is a subset of another
expression, if the expression on either side of the comparison
is an AND/OR expression, currently logic recursively traverses
down the AND/OR expression on both sides of the comparison
and if one or more child of the AND/OR is another AND/OR
expression, this recursive traversion on both sides can result
in a large number of comparisons being done in the worse case.

With the improvement of performance for flattening AND/OR
expressions (with no duplicate elimination), we can add a
pre-flattening step in case of AND/OR expressions on both sides
of the comparison such that for complex AND/OR expressions we
could save on number of comparisons being done.

Change-Id: I363f6ab2f29739da98b2daa4548cb7a45fed9bfa
Reviewed-on: https://review.couchbase.org/c/query/+/218595
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker