MB-64031 Reduce number of SubsetOf() calls
This is meant for Morpheus only.

During index selection, if an index has a WHERE clause,
the planner need to determine whether the query predicates
satisfy the index WHERE clause, which is done by calling
SubsetOf(). If the keyspace is planned multiple times,
e.g. as an inner of a nested-loop join, or hash join (if
join enumeration is not in effect), or during join
enumeration, then we can go through this WHERE clause
check multiple times. In case the index WHERE clause is
complex, this becomes time consuming and unnecessary.

Introduce a cache for both the set of indexes that have
their WHERE clause "satisfied", and the set of indexes
that have their WHERE clause "not satisfied", such that
the first time we can make this determination and put
an index in either the "satisfied" or the not satisified"
cache, keyed by the index id. Then subsequent checks can
simply get it from either cache.

The determination of whether an index's WHERE clause is
satisfied or not should only be affected by the predicates
on the keyspace, i.e. "static", and should not be affected
by join predicates, or correlation variable, etc, thus it
should be safe to keep a cache for it.

Change-Id: I03ef9dc6f2a438db4424e96b24e52c38b5999ea4
Reviewed-on: https://review.couchbase.org/c/query/+/218677
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>