MB-64058 cbq-shell: Don't strip comments from ...
... statements in files (-f/-file).  Also don't strip new-line
characters (they have meaning with single-line comments).  Any reported
error with a context will now be more easily translated to a position in
the source file (line is still statement relative, but column is
accurate).

Single-line comments starting with '#' remain and are stripped from the
statements (as they're script-only comments, not SQL++ comments).

Statements passed with -s are entirely skipped if they begin with '#'
(existing behaviour) or are a single line and start with '--'
(previously just starting with '--' was enough to exclude the whole
statement).

Change-Id: I36d265812cf635818b2a9a2f4e64e249d954d4aa
Reviewed-on: https://review.couchbase.org/c/query/+/218740
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>