MB-63947 Inherit cover transformation info from join enumeration
One more change needed for MB-63947. The previous fix for this bug
introduced two kinds of subquery covering info, saved in the builder,
one for SubqueryTerm under nested-loop join, one for SubqueryTerm
under join enumeration. The issue is if we have a SubqueryTerm under
join, but happens during join enumeration, currently we save that
in the copy of the builder used for join enumeration, but the actual
cover transformation happens later, and the builder used for join
enumeration is gone by that time. As a result the covering
transformation is not properly done. Note that this (missing) covering
transformation only affect Explain, the actual execution of the query
is not affected (in that case the SubqueryTerm is planned at execution
time).

To fix this, after join enuemration is done, check and inherit any
subquery covering information that is meant for SubqueryTerm under
join, from the builder copy used for join enumeration to the main
builder.

Change-Id: I9fe91266b3d8e3599e2ab663b2c39c5fc483b2d2
Reviewed-on: https://review.couchbase.org/c/query/+/218766
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>