MB-64086 Copy correlation references when copying formalizer
When the expression parser encounteres a "correlated " reference in
front of a subquery, it uses the formalizer to mark correlation
references for the subquery. The issue with this bug is that when
a copy of a formalizer is done, we fail to also copy the correlation
references, and thus we lost information if a copy is done. For this
bug a copy of the formalizer when formalizing the projection clause.

The fix is to check and copy correlation references when copying a
formalizer.

Change-Id: I48c04caa4d60052c4f9e1c3b9734f701055b87a4
Reviewed-on: https://review.couchbase.org/c/query/+/218964
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker