MB-64125 Revise clean-up strategy...
... for temp files.

Not all platforms can work with immediate unlinking of a file so this
changes to explicitly removing temp files when closed along with adding
a start-up clean-up to remove old temp files.

One advantage to this approach is that temp space usage is easier to
spot/monitor externally as the files can be listed/viewed.  It does mean
though that in the event of an abnormal end, they persist until the next
start-up cleans them up.

Change-Id: Iedd634df50733c6ad40dae49779a1b36945885d5
Reviewed-on: https://review.couchbase.org/c/query/+/219253
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>