MB-64062 Add request information in prepared for debugging
Copies credentials(username only), remoteAddr & userAgent fields from the
request to Prepared using the request context as an intermediate.

This way we handle passing this information to both
Prepare statements in functions and Prepare statement requests.

During a reprepare we mark the prepared as reprepared and
report "userAgent" as "creatingUserAgent".

Change-Id: I157006b646b7e4be72bbe7cec956a4ee2b030901
Reviewed-on: https://review.couchbase.org/c/query/+/219251
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>