MB-63883 Modify logic for evaluation of intersect scan with CBO
When there are multiple indexes that can be used to satisfy the
predicates on a keyspace, and optimizer statistics are available
for the keyspace, then CBO goes through a process of trying to
select the optimal number of indexes to use in a potential
intersect scan. The output of this could be a single index, or
multiple indexes in an intersect scan. The current logic for
order of consideration of the indexes is based on the
selectivities from each index, with the intention that the most
selective index should be used first, and subsequent index
considerations should also be based on selectivities. However
the issue with this logic is that the most selective index
could also be a very expensive index, and thus CBO could end
up picking an expensive index over a much cheaper index.

There could be a trade-off between the cost of an index scan vs
the selectivity offered by the index scan. The benefit of a
better selectivity is the reduction of the subsequent Fetch
cost. If the reduction of the Fetch cost does not "compensate"
for a more expensive index scan, then we should not be picking
this index.

The logic for order of consideration in this scenario is
modified to take into consideration both the cost of the index
scan as well as the selectivity of the index. The selectivity
is reflected as the cost of the subsequent Fetch of the
documents after the index scan, i.e. the more selective the
index, the less cost for fetching the documents. By combining
both the index cost and the fetch cost (instead of just
considering index selectivity), the modified logic should be
able to pick the cheapest combination of indexes to use.

MB-63883 Additional change

Although not expected, add code to deal with potential situations
where fetchCost is not available.

Change-Id: I5574051052591c270eff6550595875be9e17d083
Reviewed-on: https://review.couchbase.org/c/query/+/217910
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/219279