MB-60680 Allow ORDER BY to be in different index key order
For vector search query, if ORDER BY includes vector distance and
LIMIT is present, the indexer uses topK order where a max heap is
used for the order. This allows index order in situations where the
ORDER BY keys is not necessarily the same order as index keys. To
accommodate this, enhance the logic for index ORDER pushdown such
that we can push down index order even if the ORDER BY keys is
different than index keys.

Some of the unit tests has to skip result comparison for now
since the topK order is not complete from the indexer. We only
compare query plan for now, and will enable result comparison
once it is available from the indexer.

Change-Id: I7a92b0bc5bc4f88a5ebd5c5cdde1e7cf3faa1fa5
Reviewed-on: https://review.couchbase.org/c/query/+/219380
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>