MB-64142 Duplicate errors and warnings arrays...
... under lock when returning them to avoid concurrent access issues.

Previously we just returned the underlying arrays for iteration when
writing the response however it is possible for halting operators to
still modify them whilst this is taking place.  By locking we prevent
modification whilst we're copying and by copying we avoid possible
issues if subsequent modification relocates the array memory and the
"old" memory is reused before the iteration is complete.

Added "commands" to LOGLEVEL=debug filters indicated with a leading '!'.
'!+dumpcore' turns on core dumping and '!-dumpcore' turns off core
dumping.  Filter commands can be combined with other filters.  Core
dumping remains enabled until explicitly disabled, even if the LOGLEVEL
is to to a level other than debug.

Change-Id: Ied16b804c796b2091f643f00481267b22d5ef9bf
Reviewed-on: https://review.couchbase.org/c/query/+/219549
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker