MB-60680 Allow covering in BHive index if actual_vector ...      ... is set to true but reranking is used
The BHive index has the capability to "rerank", i.e. after the
candidates are retrieved, it can then get the actual vectors
and use that to recompute distances (the index uses a compressed
vector) which is more accurate. Current code logic disables
covering if "actual_vector" parameter is set to true in the ANN
expression.

Enhance the logic such that if we detect a BHive index with
"rerank" capability, allows covering if the "actual_vector"
parameter is set to true.

Change-Id: I95c8c9e475fbd88beca6b8bc3e7b20ae97af8e60
Reviewed-on: https://review.couchbase.org/c/query/+/219673
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>