MB-63252 MB-63659 Unit-Test code changes
Modify runtest.sh/runtests.sh:
* prompt for natural_cred and natural_orgid
  and export them. They can later be read as part of
  constructing a natural language request
  (for USING AI statement tests)
* add test_cases/natural to -s skiptests in runtests.sh

Add a new field "generated_stmt" in the testcase.
It is set to true when we have a Natural Language request
and a statement is generated successfully.

Reason for matching with a boolean is that for
a Natural request we can have more than one statement
that may satisfy the expectation and we can't
always be sure the generated statement would be same.
One thing we can be sure of is the results(incase of a select).

Move ProcessNatural() to *server.BaseRequest so we can
use it in the test code as well.

Change-Id: If71c9afe2748d0d224f466e4750bf4bc587780d6
Reviewed-on: https://review.couchbase.org/c/query/+/219696
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>