MB-63507 Modify previous fix
The previous fix for this bug introduced a new measurement
for the number of index keys that matches EQ/IN predicates
in index condition. This new measurement is then used to
further differentiate two similar indexes and avoid
potential intersect scans. The change however has an
unintentional side effect and sometimes can result in an
intersect scan where previously it would not.

We now fine-tune the new measurement check in deciding
whether one index is better than another. Two aspects:
 - add number of index condition matches to consideration
   since we are looking for index condition matches anyway
 - add another check for index condition subset as well,
   such that we don't accidentially prefer an index that
   would have been discarded if the indexes were
   considered in reverse order

Change-Id: I38db14f80ec2b3620e4e83872af777fba58e1a63
Reviewed-on: https://review.couchbase.org/c/query/+/219774
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>