MB-60680 Add consideration of vector index key in index costing
For vector index key the planner generates a _WHOLE_SPANS and the
indexer relies on a separate parameter for vector-related information.
For costing purpose, we get number of centroids from the index, and
nprobes from either the ANN expression or the index definition, and
estimate a selectivity for the vector index key.

Note that CBO is currently still disabled. The indexer is not
returning valid number of centroids yet.

Change-Id: Id2bb09d53daa9077621c1bc3f0c886f5db2f98d2
Reviewed-on: https://review.couchbase.org/c/query/+/219903
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>