MB-64352 Use proper keyspace name when getting Fetch cost
This is an extension of the fix for MB-63883.
The fix for MB-63883 added Fetch cost into consideration when
CBO chooses intersect scans. For this bug, the index has several
intersect spans underneath a union span. To add Fetch cost, the
previous fix for MB-63883 calls the function for calculating
Fetch cost with the "alias" of the keyspace. This does not work
properly, the function needs the actual keyspace name in order
to properly get to the optimizer stats for that keyspace.

The fix is to use proper keyspace name when getting Fetch cost.
This requires the keyspace name being passed through the
function calls (extra parameters to a few functions).

Change-Id: I7ba04e83262c8351b1ab55e0a71eccd995d0d9be
Reviewed-on: https://review.couchbase.org/c/query/+/219927
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>