MB-62471: [7/n] Do not collect change information in CBO statistics...
If:
1. AUS is disabled
and/or
2. The histogram is internally created by the Advisor

A check is made if AUS is disabled, and if it is, the change
information is not collected and stored in the histogram.

A side-effect of this is that, if AUS is disabled mid-way
during a scheduled AUS window, the change information will not
be collected for any CBO stats that are updated after this disablement.

Change-Id: I37143223a79fd2c1028da984e3e7186e021fc5ed
Reviewed-on: https://review.couchbase.org/c/query/+/219958
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>