MB-64369 Do not pushdown volatile filters to base keyspace
When a query involves a join, then we attempt to pushdown
filters on a single keyspace to below the join, in an attempt
to reduce number input feeding into the join. Normally when
this is done we remove the pushed down filter from the final
filtering step. In this bug, we have a volatile filter which
uses the RANDOM() function, and volatile filter expressions
always return FALSE for EquivalentTo() function call, which
is used for removing the filter expression. As a result, the
filter expression is not removed from final filtering step
resulting in double application of the volatile filter
expression evaluation.

The fix is to skip volatile filter expressions for pushing
down to base keyspace.

Change-Id: I0eb8ff572dc30708736249d35c1889186c7e631a
Reviewed-on: https://review.couchbase.org/c/query/+/219982
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>