MB-64408 Remove keyspace reference from correlation map ...    ... when the keyspace is in scope (and no longer correlated)
In this bug we have a CTE definition that involves a nested ANY
clause. The formalization process creates a new "scope" when it
see an ANY clause, such that it can incorporate the binding
variables as allowed references. Due to this extra scope, a
reference to the keyspace alias in the query is interpreted as
correlated since it is defined one-level above. As a result the
keyspace reference is added to the correlation map of the
formalizer. Prior to the fix for MB-64086, when we make a copy
of the formalizer we did not copy over the correlation map and
thus that information just got lost. With the fix for MB-64086,
we now have the proper correlation map copied over, however
when we pop the extra scope for the ANY clause, the keyspace
reference is now in scope, and the formalizer properly recognized
it as "allowed", however the keyspace reference is still in the
correlation map of the formalizer, which results in sanity
check failures later on.

The fix is to add logic to the function CheckCorrelated(), such
that if an identifier is "allowed", and it is in the correlation
map, remove it from the correlation map since it is no longer
correlated.

Change-Id: I4766329edd46a1711d4e2382b2b4dc9cd66c778a
Reviewed-on: https://review.couchbase.org/c/query/+/220128
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>