MB-60680 Keep track of number of "sargable" include columns ...    ... in index entry
The planner uses an index entry to keep information about a usable
index for the query. The index entry currently keeps the include
columns defined for the index, however it does not know whether
any of the include columns are "sargable", i.e. there are query
predicate that can be used to filter on an include column.

Add number of sargable include columns to an index entry. This
information is derived when sargability information is derived
for index keys, and consideration of include columns is added
such that they are considered at the same time as the regular
index keys.

When RBO compares two vector indexes, add criteria to compare
number of sargable include columns.

Change-Id: I1bd9c748a450b231e29c0ca6266d082e71514f45
Reviewed-on: https://review.couchbase.org/c/query/+/220183
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>