MB-64423 Do not cache field name value if it contains function variable
The Field expression has a cache for the 2nd part (field name) if it
determines that the field name is "static". This can be problematic
for function variables since they are marked with IDENT_IS_STATIC_VAR
flag and thus considered "static". With recursive function involcation
we are getting incorrect values due to this cache.

The fix is to add a new expression interface function to detect
references to static variables (WITH alias and function variables)
and do not cache the field name in such cases.

Change-Id: I8efa2bf135311a2bfacb275640bf29cd0ec4f445
Reviewed-on: https://review.couchbase.org/c/query/+/220191
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker