MB-64409: Fix error when parsing inline function's expression string...
The issue is due to the expression.UserDefinedFunction.String() method
returning the full name of the UDF function without backticks.

Example:
CREATE FUNCTION test1() { `func-1`()};
The execution of such a function would return a parsing error.

Issues caused by this:

1. When executing an inline UDF, the string of the body's expression is parsed.

If the body contains a reference to a function whose full name contains
parts that require being escaped by backticks, the parsing fails.

2.  The unmarshalling of prepared statements that refer to such UDFs
would fail due parsing failure of the UDF's expression string.

And hence would only execute on the node the prepared statement
was created in.

Fix:
The fix is to return the full name of the UDF function with backticks when calling the expression.UserDefinedFunction.String() method.

Change-Id: I9287fbd4d6cf2722764fe2b6da7ebe07538024f1
Reviewed-on: https://review.couchbase.org/c/query/+/220195
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>