MB-64442 Exclude static variables when appropriate
We use the Static() function in an expression for a number
of purposes. The Static() function considers a WITH alias
as well as a function variable as static, the intent there
was that the value of these are external to the query
being executed and thus can be considered as a constant
while the query is executing. However, there are situations
where the scope of these determinations can go beyond a
single query execution. In these situations we should not
treat them as constants. An example of this is MB-64423.

Introduce a new expression interface function StaticNoVariable(),
which is similar to the existing Static() function, except
it does not include static variables. Then wherever we
should exclude static variables from consideration we
should use this new function instead of the original
Static() function.

Change-Id: I99fb9318eed7821eca2cc6a88dace63d444fef6e
Reviewed-on: https://review.couchbase.org/c/query/+/220231
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>